Split mentorController
